requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
apscheduler==3.10.4
sqlite3
requests==2.31.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-dateutil==2.8.2
pytz==2023.3